fix(activerecord): move buildColumnSerializer into attribute-methods/serialization.ts (api:compare attribution)
